Updat module crawler; refactor code
