+++ fix_scripts/fix_plan.py
@@ -2,7 +2,8 @@ import sqlite3
 from plan.util import get_degree_info
 
 # Database path (adjust this if necessary)
-DB_PATH = './res/data.db'
+DB_PATH = "./res/data.db"
+
 
 # Step 1: Add new columns `degree` and `degreeType` to the `plan` table
 def add_columns():
@@ -16,9 +17,10 @@ def add_columns():
     except sqlite3.OperationalError as e:
         print(f"Error occurred while adding columns: {e}")
         # If columns already exist, you can catch and continue
-    
+
     conn.commit()
     conn.close()
+
 
 # Step 2: Update existing rows with `degree` and `degreeType` information
 def update_plans_with_degree_info():
@@ -34,18 +36,22 @@ def update_plans_with_degree_info():
 
         # Fetch degree information using the `get_degree_info` function
         degree_info = get_degree_info(title)
-        degree = degree_info['degree']
-        degree_type = degree_info['degreeType']
+        degree = degree_info["degree"]
+        degree_type = degree_info["degreeType"]
 
         # Update the row with the degree and degreeType info
-        cursor.execute('''
+        cursor.execute(
+            """
             UPDATE plan
             SET degree = ?, degreeType = ?
             WHERE year = ? AND campus = ? AND academicPlanCode = ?
-        ''', (degree, degree_type, year, campus, academicPlanCode))
+        """,
+            (degree, degree_type, year, campus, academicPlanCode),
+        )
 
     conn.commit()
     conn.close()
+
 
 def main():
     # Step 1: Alter the table to add the new columns
@@ -55,6 +61,7 @@ def main():
     update_plans_with_degree_info()
 
     print("Database updated successfully.")
+
 
 if __name__ == "__main__":
     main()
