+++ module/config.py
@@ -5,9 +5,7 @@ SCHOOLS_PATH = "./res/schools.json"
 MODULE_BRIEF_PATH = "./res/module_brief.json"
 DB_PATH = "./res/data.db"
 COURSE_TABLE_NAME = "course"
-THREADS = (
-    9  # Could be more if you have a powerful computer and stable Internet
-)
+THREADS = 9  # Could be more if you have a powerful computer and stable Internet
 HEADLESS = True  # Set to False for testing
 # Define the path to chromedriver.exe
 DRIVER_PATH = "C:\\Users\\eric\\apps\\chromedriver.exe"
